+++ pyatmo/Exceptions.py
