+++ pyatmo/HomeCoach.py
@@ -1,5 +1,5 @@
-from . import _BASE_URL
-from .WeatherStation import WeatherStationData
+from pyatmo import _BASE_URL
+from pyatmo.WeatherStation import WeatherStationData
 
 _GETHOMECOACHDATA_REQ = _BASE_URL + "api/gethomecoachsdata"
