+++ tests/test_pyatmo_camera.py
@@ -4,9 +4,8 @@ import json
 import pytest
 from freezegun import freeze_time
 
-import smart_home.Camera
+import pyatmo.Camera
 
-from .conftest import does_not_raise
 
 INVALID_NAME = "InvalidName"
 
@@ -43,7 +42,7 @@ def test_CameraData_homeById(cameraHomeData, hid, expected):
 )
 def test_CameraData_homeByName(cameraHomeData, name, expected):
     if name == INVALID_NAME:
-        with pytest.raises(smart_home.Exceptions.InvalidHome):
+        with pytest.raises(pyatmo.Exceptions.InvalidHome):
             assert cameraHomeData.homeByName(name)
     else:
         assert cameraHomeData.homeByName(name)["id"] == expected
@@ -147,11 +146,11 @@ def test_CameraData_cameraUrls_disconnected(auth, requests_mock):
     with open("fixtures/camera_home_data_disconnected.json") as f:
         json_fixture = json.load(f)
     requests_mock.post(
-        smart_home.Camera._GETHOMEDATA_REQ,
+        pyatmo.Camera._GETHOMEDATA_REQ,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
-    camera_data = smart_home.Camera.CameraData(auth)
+    camera_data = pyatmo.Camera.CameraData(auth)
     assert camera_data.cameraUrls() == (None, None)
 
 
@@ -169,7 +168,7 @@ def test_CameraData_cameraUrls_disconnected(auth, requests_mock):
 )
 def test_CameraData_personsAtHome(cameraHomeData, home, expected):
     if home == INVALID_NAME:
-        with pytest.raises(smart_home.Exceptions.InvalidHome):
+        with pytest.raises(pyatmo.Exceptions.InvalidHome):
             assert cameraHomeData.personsAtHome(home)
     else:
         assert cameraHomeData.personsAtHome(home) == expected
@@ -334,5 +333,5 @@ def test_CameraData_getHomeName(cameraHomeData):
 def test_CameraData_gethomeId(cameraHomeData):
     assert cameraHomeData.gethomeId() == "91763b24c43d3e344f424e8b"
     assert cameraHomeData.gethomeId("MYHOME") == "91763b24c43d3e344f424e8b"
-    with pytest.raises(smart_home.Camera.InvalidHome):
+    with pytest.raises(pyatmo.Camera.InvalidHome):
         assert cameraHomeData.gethomeId("InvalidName")
