+++ tests/test_pyatmo_homecoach.py
@@ -3,7 +3,7 @@ import json
 
 import pytest
 
-import smart_home.HomeCoach
+import pyatmo.HomeCoach
 
 
 def test_HomeCoachData(homeCoachData):
@@ -88,9 +88,9 @@ def test_HomeCoachData_no_devices(auth, requests_mock):
     with open("fixtures/home_coach_no_devices.json") as f:
         json_fixture = json.load(f)
     requests_mock.post(
-        smart_home.HomeCoach._GETHOMECOACHDATA_REQ,
+        pyatmo.HomeCoach._GETHOMECOACHDATA_REQ,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
-    with pytest.raises(smart_home.WeatherStation.NoDevice):
-        assert smart_home.HomeCoach.HomeCoachData(auth)
+    with pytest.raises(pyatmo.WeatherStation.NoDevice):
+        assert pyatmo.HomeCoach.HomeCoachData(auth)
