+++ tests/test_pyatmo_weatherstation.py
@@ -4,7 +4,7 @@ import json
 import pytest
 from freezegun import freeze_time
 
-import smart_home.WeatherStation
+import pyatmo.WeatherStation
 
 from .conftest import does_not_raise
 
@@ -14,33 +14,33 @@ def test_WeatherStationData(weatherStationData):
 
 
 def test_WeatherStationData_no_response(auth, requests_mock):
-    requests_mock.post(smart_home.WeatherStation._GETSTATIONDATA_REQ, text="None")
-    with pytest.raises(smart_home.WeatherStation.NoDevice):
-        assert smart_home.WeatherStation.WeatherStationData(auth)
+    requests_mock.post(pyatmo.WeatherStation._GETSTATIONDATA_REQ, text="None")
+    with pytest.raises(pyatmo.WeatherStation.NoDevice):
+        assert pyatmo.WeatherStation.WeatherStationData(auth)
 
 
 def test_WeatherStationData_no_body(auth, requests_mock):
     with open("fixtures/status_ok.json") as f:
         json_fixture = json.load(f)
     requests_mock.post(
-        smart_home.WeatherStation._GETSTATIONDATA_REQ,
+        pyatmo.WeatherStation._GETSTATIONDATA_REQ,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
-    with pytest.raises(smart_home.WeatherStation.NoDevice):
-        assert smart_home.WeatherStation.WeatherStationData(auth)
+    with pytest.raises(pyatmo.WeatherStation.NoDevice):
+        assert pyatmo.WeatherStation.WeatherStationData(auth)
 
 
 def test_WeatherStationData_no_data(auth, requests_mock):
     with open("fixtures/home_data_empty.json") as f:
         json_fixture = json.load(f)
     requests_mock.post(
-        smart_home.WeatherStation._GETSTATIONDATA_REQ,
+        pyatmo.WeatherStation._GETSTATIONDATA_REQ,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
-    with pytest.raises(smart_home.WeatherStation.NoDevice):
-        assert smart_home.WeatherStation.WeatherStationData(auth)
+    with pytest.raises(pyatmo.WeatherStation.NoDevice):
+        assert pyatmo.WeatherStation.WeatherStationData(auth)
 
 
 @pytest.mark.parametrize(
@@ -519,7 +519,7 @@ def test_WeatherStationData_getMeasure(
     with open("fixtures/weatherstation_measure.json") as f:
         json_fixture = json.load(f)
     requests_mock.post(
-        smart_home.WeatherStation._GETMEASURE_REQ,
+        pyatmo.WeatherStation._GETMEASURE_REQ,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
